types refactoring
